Fisys/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
qrcode
pillow
python-multipart
netifaces
requests
websockets
passlib
bcrypt
itsdangerous
python-dotenv
psycopg2-binary